+++ src/main/java/dev/shteryu/star_wars/repository/PeoplePagingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import dev.shteryu.star_wars.models.People;
 
 @Repository
-public interface PeoplePagingRepository extends PagingAndSortingRepository<People, Integer>{
+public interface PeoplePagingRepository extends PagingAndSortingRepository<People, Long>{
     
 }
