+++ src/main/java/dev/shteryu/star_wars/repository/PlanetPagingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import dev.shteryu.star_wars.models.Planets;
 
 @Repository
-public interface PlanetPagingRepository extends PagingAndSortingRepository<Planets, Integer>{
+public interface PlanetPagingRepository extends PagingAndSortingRepository<Planets, Long>{
     
 }
