+++ src/main/java/dev/shteryu/star_wars/repository/StarshipPagingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import dev.shteryu.star_wars.models.Starships;
 
 @Repository
-public interface StarshipPagingRepository extends PagingAndSortingRepository<Starships, Integer>{
+public interface StarshipPagingRepository extends PagingAndSortingRepository<Starships, Long>{
     
 }
