+++ src/main/java/dev/shteryu/star_wars/repository/VehiclePagingRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import dev.shteryu.star_wars.models.Vehicles;
 
 @Repository
-public interface VehiclePagingRepository extends PagingAndSortingRepository<Vehicles, Integer>{
+public interface VehiclePagingRepository extends PagingAndSortingRepository<Vehicles, Long>{
     
 }
