+++ src/main/java/dev/shteryu/star_wars/web/dto/FilmPeopleGetResponse.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class FilmPeopleGetResponse {
 
-    private Set<Integer> filmPeopleIds;
+    private Set<Long> filmPeopleIds;
 
 }
