+++ src/main/java/dev/shteryu/star_wars/web/dto/FilmPlanetsUpsertRequest.java
@@ -14,6 +14,6 @@ import lombok.NoArgsConstructor;
 public class FilmPlanetsUpsertRequest {
 
     @NotNull
-    private Set<Integer> filmPlanetsIds;
+    private Set<Long> filmPlanetsIds;
 
 }
